+++ app/widgets/hue/src/Hue.kt
@@ -2,35 +2,12 @@
 
 package hue
 
-import de.jensd.fx.glyphs.GlyphIcons
-import hue.HueBulbConfPowerOn.custom
-import hue.HueBulbConfPowerOn.unknown
-import hue.HueGroupType.LightGroup
-import hue.HueGroupType.Lightsource
-import hue.HueGroupType.Luminaire
-import hue.HueGroupType.Room
-import hue.HueGroupType.Zone
-import hue.HueSceneType.GroupScene
-import hue.HueSceneType.LightScene
-import io.ktor.client.HttpClient
-import io.ktor.client.engine.cio.CIO
-import io.ktor.client.request.HttpRequestBuilder
-import io.ktor.client.request.delete
-import io.ktor.client.request.get
-import io.ktor.client.request.post
-import io.ktor.client.request.put
-import io.ktor.client.request.setBody
-import io.ktor.client.statement.bodyAsText
-import io.ktor.http.HttpStatusCode.Companion.OK
-import io.ktor.http.HttpStatusCode.Companion.SwitchingProtocols
 import java.awt.Color.HSBtoRGB
 import javafx.event.EventHandler
 import javafx.geometry.Pos.TOP_CENTER
 import javafx.geometry.Pos.TOP_LEFT
 import javafx.geometry.Side.RIGHT
-import javafx.scene.Node
 import javafx.scene.control.ContextMenu
-import javafx.scene.control.Label
 import javafx.scene.control.ScrollPane.ScrollBarPolicy.AS_NEEDED
 import javafx.scene.control.ScrollPane.ScrollBarPolicy.NEVER
 import javafx.scene.control.TitledPane
@@ -48,20 +25,14 @@ import javafx.scene.input.MouseEvent.MOUSE_PRESSED
 import javafx.scene.input.MouseEvent.MOUSE_RELEASED
 import javafx.scene.layout.Priority.ALWAYS
 import javafx.scene.layout.StackPane
-import javafx.scene.layout.VBox
 import javafx.scene.paint.Color
 import javafx.scene.paint.Color.TRANSPARENT
 import javafx.scene.shape.Circle
-import javafx.scene.text.TextAlignment
 import kotlin.math.PI
 import kotlin.math.atan2
-import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
-import kotlinx.coroutines.async
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import mu.KLogging
-import sp.it.pl.core.InfoUi
 import sp.it.pl.layout.Widget
 import sp.it.pl.layout.WidgetCompanion
 import sp.it.pl.layout.appProperty
@@ -70,8 +41,6 @@ import sp.it.pl.main.APP
 import sp.it.pl.main.IconFA
 import sp.it.pl.main.IconMA
 import sp.it.pl.main.IconMD
-import sp.it.pl.main.IconOC
-import sp.it.pl.main.IconWH
 import sp.it.pl.main.Key
 import sp.it.pl.main.WidgetTags.IOT
 import sp.it.pl.main.WidgetTags.VISUALISATION
@@ -79,9 +48,25 @@ import sp.it.pl.main.configure
 import sp.it.pl.main.emScaled
 import sp.it.pl.main.showFloating
 import sp.it.pl.main.textColon
-import sp.it.pl.main.toUi
-import sp.it.pl.plugin.impl.SpeechRecognition
-import sp.it.pl.plugin.impl.SpeechRecognition.SpeakHandler
+import sp.it.pl.plugin.Hue
+import sp.it.pl.plugin.HueBulb
+import sp.it.pl.plugin.HueBulbConfPowerOn.custom
+import sp.it.pl.plugin.HueBulbConfPowerOn.unknown
+import sp.it.pl.plugin.HueBulbId
+import sp.it.pl.plugin.HueBulbState
+import sp.it.pl.plugin.HueBulbStateEditLight
+import sp.it.pl.plugin.HueCell
+import sp.it.pl.plugin.HueCellNode
+import sp.it.pl.plugin.HueGroup
+import sp.it.pl.plugin.HueGroupCreate
+import sp.it.pl.plugin.HueGroupId
+import sp.it.pl.plugin.HueGroupType.LightGroup
+import sp.it.pl.plugin.HueGroupType.Lightsource
+import sp.it.pl.plugin.HueGroupType.Luminaire
+import sp.it.pl.plugin.HueGroupType.Room
+import sp.it.pl.plugin.HueGroupType.Zone
+import sp.it.pl.plugin.HueIcon
+import sp.it.pl.plugin.HueSceneCreate
 import sp.it.pl.ui.objects.form.Form.Companion.form
 import sp.it.pl.ui.objects.form.Validated
 import sp.it.pl.ui.objects.icon.Icon
@@ -92,9 +77,7 @@ import sp.it.util.access.v
 import sp.it.util.async.FX_LATER
 import sp.it.util.async.coroutine.FX
 import sp.it.util.async.coroutine.asFut
-import sp.it.util.async.coroutine.runSuspendingFx
 import sp.it.util.async.future.Fut
-import sp.it.util.async.runFX
 import sp.it.util.collections.setTo
 import sp.it.util.conf.ConfigurableBase
 import sp.it.util.conf.EditMode.NONE
@@ -106,26 +89,16 @@ import sp.it.util.conf.cvn
 import sp.it.util.conf.def
 import sp.it.util.conf.lengthMax
 import sp.it.util.conf.uiConverterElement
-import sp.it.util.dev.fail
-import sp.it.util.dev.failIf
 import sp.it.util.file.div
-import sp.it.util.file.json.JsNull
-import sp.it.util.file.json.JsObject
-import sp.it.util.file.json.JsString
-import sp.it.util.file.json.JsValue
-import sp.it.util.file.json.div
-import sp.it.util.file.json.toPrettyS
 import sp.it.util.functional.Try
 import sp.it.util.functional.asIf
 import sp.it.util.functional.asIs
 import sp.it.util.functional.ifNotNull
+import sp.it.util.functional.ifNull
 import sp.it.util.functional.net
-import sp.it.util.functional.orNull
-import sp.it.util.functional.runTry
 import sp.it.util.functional.toUnit
 import sp.it.util.math.clip
 import sp.it.util.math.min
-import sp.it.util.reactive.Subscription
 import sp.it.util.reactive.Suppressor
 import sp.it.util.reactive.attachTrue
 import sp.it.util.reactive.consumeScrolling
@@ -138,11 +111,8 @@ import sp.it.util.reactive.sync1IfInScene
 import sp.it.util.reactive.syncFrom
 import sp.it.util.system.browse
 import sp.it.util.text.capitalLower
-import sp.it.util.text.equalsNc
 import sp.it.util.text.keys
 import sp.it.util.text.nameUi
-import sp.it.util.text.split3
-import sp.it.util.type.atomic
 import sp.it.util.ui.alpha
 import sp.it.util.ui.center
 import sp.it.util.ui.centre
@@ -151,7 +121,6 @@ import sp.it.util.ui.flowPane
 import sp.it.util.ui.hBox
 import sp.it.util.ui.label
 import sp.it.util.ui.lay
-import sp.it.util.ui.lookupId
 import sp.it.util.ui.prefSize
 import sp.it.util.ui.pseudoClassChanged
 import sp.it.util.ui.scrollPane
@@ -161,14 +130,12 @@ import sp.it.util.ui.unitCircleDegP
 import sp.it.util.ui.vBox
 import sp.it.util.ui.x
 import sp.it.util.ui.xy
-import sp.it.util.units.seconds
 import sp.it.util.units.uri
 import sp.it.util.units.version
 import sp.it.util.units.year
 
 class Hue(widget: Widget): SimpleController(widget) {
 
-   val client = HttpClient(CIO).apply { onClose += ::close }
    val hueBulbCells = HashMap<HueBulbId, HueCell<HueBulb>>()
    val hueBulbGroupCells = HashMap<HueGroupId, HueCell<HueGroup>>()
    var selectedGroupId: HueGroupId? = null
@@ -179,160 +146,30 @@ class Hue(widget: Widget): SimpleController(widget) {
    val groupsPane = flowPane(10.emScaled, 10.emScaled)
    val scenesPane = flowPane(10.emScaled, 10.emScaled)
    val sensorsPane = flowPane(10.emScaled, 10.emScaled)
-   val speechHandlers = mutableListOf<SpeakHandler>()
-
-   private val hueBridge = object: CoroutineScope by scope {
-      val hueBridgeUserDevice = "spit-player"
-      lateinit var ip: String
-      lateinit var apiKey: String
-      lateinit var apiVersion: KotlinVersion
-      lateinit var url: String
-
-      suspend fun init() {
-         ip = hueBridgeIp.validIpOrNull() ?: ip() ?: fail { "Unable to obtain Phillips Hue bridge ip. Make sure it is turned on and connected to the network." }
-         hueBridgeIp = ip
-         apiKey = if (isAuthorizedApiKey(ip, hueBridgeApiKey)) hueBridgeApiKey else createApiKey(ip)
-         hueBridgeApiKey = apiKey
-         apiVersion = apiVersion("http://$ip/api/$apiKey").split3(".").net { (major, minor, patch) -> KotlinVersion(major.toInt(), minor.toInt(), patch.toIntOrNull() ?: 0) }
-         url = "http://$ip/api/$apiKey"
-      }
-
-      private suspend fun String.validIpOrNull(): String? {
-         val isValid = runTry { client.get("http://${this@validIpOrNull}").status }.orNull()==OK
-         return this@validIpOrNull.takeIf { isValid }
-      }
-
-      private suspend fun ip(): String? {
-         val response = client.get("https://discovery.meethue.com/").bodyAsText()
-         return (response.parseToJson().asJsArray()/0/"internalipaddress")?.asJsStringValue()
-      }
-
-      private suspend fun isAuthorizedApiKey(ip: String, apiKey: String): Boolean {
-         val response = client.get("http://$ip/api/$apiKey/lights")
-         failIf(response.status!=OK) { "Failed to check if hue bridge user $apiKey exists" }
-         return runTry { (response.bodyAsText().parseToJson()/0/"error") }.isError
-      }
-
-      private suspend fun createApiKey(ip: String): String {
-         logger.info { "Creating Phillips Hue bridge user" }
-
-         val pressCheckDuration = 30.seconds
-         var pressed by atomic(false)
-         runFX {
-            showFloating("Hue bridge linking") { popup ->
+   val huePlugin = APP.plugins.plugin<Hue>().asValue()
+   val hueBridge: Hue.HueBridge?
+      get() {
+         huePlugin.value.ifNull {
+            showFloating("Hue plugin") { popup ->
                vBox {
-                  lay += text("Please press the Phillips Hue bridge link button within ${pressCheckDuration.toUi()}")
-                  lay += Icon(IconFA.CHECK).run {
-                     onClickDo {
-                        popup.hide()
-                        pressed = true
+                  val hue = APP.plugins.getRaw<Hue>()
+                  if (hue==null) {
+                     lay += text("Hue plugin not available. Check whether the plugin is installed.")
+                  } else {
+                     lay += text("Hue plugin not running.")
+                     lay += Icon(IconMD.RUN).run {
+                        onClickDo {
+                           popup.hide()
+                           hue.start()
+                        }
+                        withText(RIGHT, "Start plugin...")
                      }
-                     withText(RIGHT, "I pressed it, continue...")
                   }
                }
             }
          }
-         repeat(pressCheckDuration.toSeconds().toInt() + 5) {
-            if (!pressed) delay(1000)
-         }
-
-         val userCreated = client.post("http://$ip/api") {
-            setBody(JsObject("devicetype" to JsString(hueBridgeUserDevice)).toPrettyS())
-         }
-         failIf(userCreated.status==SwitchingProtocols) { "Failed to create Phillips Hue bridge user: Link button not pressed" }
-         failIf(userCreated.status!=OK) { "Failed to create Phillips Hue bridge user" }
-         val apiKey = runTry { (userCreated.bodyAsText().parseToJson()/0/"success"/"username")?.asJsStringValue() }.orThrow
-         return apiKey ?: fail { "Failed to obtain user Phillips Hue bridge api key" }
+         return huePlugin.value?.hueBridge
       }
-
-      private suspend fun apiVersion(url: String): String {
-         val response = client.getText("$url/config")
-         return response.parseToJson().asJsObject().value["apiversion"]?.asJsString()?.value ?: fail { "Could not obtain api version" }
-      }
-
-      suspend fun bulbs(): List<HueBulb> = client.getText("$url/lights")
-         .parseToJson().asJsObject().value.map { (id, bulbJs) -> bulbJs.to<HueBulb>().copy(id = id) }
-
-      suspend fun groups(): List<HueGroup> = client.getText("$url/groups")
-         .parseToJson().asJsObject().value.map { (id, bulbJs) -> bulbJs.to<HueGroup>().copy(id = id) }
-
-      suspend fun bulbsAndGroups(): Pair<List<HueBulb>, List<HueGroup>> = (bulbs() to groups()).net { (bulbs, groups) ->
-         bulbs to groups + HueGroup("0", "All", listOf(), HueGroupState(bulbs.all { it.state.on }, bulbs.any { it.state.on }), null)
-      }
-
-      suspend fun scenes(): List<HueScene> = client.getText("$url/scenes")
-         .parseToJson().asJsObject().value.map { (id, sceneJs) -> sceneJs.to<HueScene>().copy(id = id) }
-
-      suspend fun sensors(): List<HueSensor> =client.getText("$url/sensors")
-         .parseToJson().asJsObject().value.map { (id, sceneJs) -> sceneJs.to<HueSensor>().copy(id = id) }
-
-      fun renameBulb(bulb: HueBulbId, name: String) = runSuspendingFx {
-         client.put("$url/lights/$bulb") {
-            setBody("""{"name": "$name"}""")
-         }
-      }
-
-      fun changePowerOn(bulb: HueBulbId, powerOn: HueBulbConfPowerOn) = runSuspendingFx {
-         client.putText("$url/lights/$bulb/config") {
-            setBody("""{ "startup": {"mode": "$powerOn"} }""")
-         }
-      }
-
-      fun toggleBulb(bulb: HueBulbId) = runSuspendingFx {
-         val response = client.getText("$url/lights/$bulb")
-         val on = response.parseToJson().to<HueBulb>().state.on
-         client.putText("$url/lights/$bulb/state") {
-            setBody(HueBulbStateEditOn(!on).toJson().toPrettyS())
-         }
-      }
-
-      fun toggleBulbGroup(group: HueGroupId) = runSuspendingFx {
-         val response = client.getText("$url/groups/$group")
-         val allOn = response.parseToJson().to<HueGroup>().copy(id = group).state.all_on
-         client.putText("$url/groups/$group/action") {
-            setBody(HueBulbStateEditOn(!allOn).toJson().toPrettyS())
-         }
-      }
-
-      fun applyBulbLight(bulb: HueBulbId, state: HueBulbStateEditLight) = runSuspendingFx {
-         client.putText("$url/lights/$bulb/state") {
-            setBody(state.toJson().asJsObject().withoutNullValues().toPrettyS())
-         }
-      }
-
-      fun applyBulbGroupLight(group: HueGroupId, state: HueBulbStateEditLight) = runSuspendingFx {
-         client.putText("$url/groups/$group/action") {
-            setBody(state.toJson().asJsObject().withoutNullValues().toPrettyS())
-         }
-      }
-
-      fun applyScene(scene: HueScene) = runSuspendingFx {
-         client.putText("$url/groups/0/action") {
-            setBody(JsObject("scene" to JsString(scene.id)).toPrettyS())
-         }
-      }
-
-      fun createGroup(group: HueGroupCreate) = runSuspendingFx {
-         client.postText("$url/groups") {
-            setBody(group.toJson().toPrettyS())
-         }
-      }
-
-      fun createScene(scene: HueSceneCreate) = async {
-         client.postText("$url/scenes") {
-            setBody(scene.toJson().asJsObject().withoutNullValues().toPrettyS())
-         }
-      }
-
-      fun deleteGroup(group: HueGroupId) = runSuspendingFx {
-         client.deleteText("$url/groups/$group")
-      }
-
-      fun deleteScene(group: HueSceneId) = runSuspendingFx {
-         client.deleteText("$url/scenes/$group")
-      }
-
-   }
 
    private val infoPane = vBox(30.emScaled)
    private val devicePane = vBox(0, TOP_LEFT)
@@ -417,8 +254,8 @@ class Hue(widget: Widget): SimpleController(widget) {
       fun applyToSelected(bri: Int?, hue: Int?, sat: Int?) {
          avoidApplying.suppressed {
             val state = HueBulbStateEditLight(bri, hue, sat)
-            selectedBulbId.ifNotNull { hueBridge.applyBulbLight(it, state) }
-            selectedGroupId.ifNotNull { hueBridge.applyBulbGroupLight(it, state) }
+            selectedBulbId.ifNotNull { hueBridge?.applyBulbLight(it, state) }
+            selectedGroupId.ifNotNull { hueBridge?.applyBulbGroupLight(it, state) }
          }
       }
    }
@@ -464,19 +301,12 @@ class Hue(widget: Widget): SimpleController(widget) {
          }
       }
 
-      onClose += { APP.plugins.use<SpeechRecognition> { it.handlers -= speechHandlers } }
-      APP.plugins.attachWhile<SpeechRecognition> {
-         it.handlers += speechHandlers
-         Subscription { it.handlers -= speechHandlers }
-      } on onClose
-
       root.sync1IfInScene { refresh() } on onClose
-      onClose += { client.close() }
    }
 
-   fun Fut<*>.thenRefresh() = then(FX_LATER) { refresh().toUnit() }.toUnit()
+   fun Fut<*>?.thenRefresh() = this?.then(FX_LATER) { refresh().toUnit() }.toUnit()
 
-   fun linkBridge(): Job = scope.launch(FX) { hueBridge.init() }
+   fun linkBridge(): Job = scope.launch(FX) { hueBridge?.init() }
 
    fun refresh(): Job = scope.launch(FX) {
 
@@ -499,28 +329,15 @@ class Hue(widget: Widget): SimpleController(widget) {
          infoPane.lay -= devicePane
       }
 
-      hueBridge.init()
-      hueBridge.bulbsAndGroups().net { (bulbs, groups) ->
-         APP.plugins.use<SpeechRecognition> {
-            it.handlers -= speechHandlers
-            speechHandlers.clear()
-            speechHandlers += SpeakHandler("Turn all lights on/off",   "lights on|off") { text, _ -> if ("lights on" in text || "lights off" in text) hueBridge.toggleBulbGroup("0").thenRefresh() }
-            speechHandlers += SpeakHandler("Turn group lights on/off", "lights \$group-name on|off") { text, _ ->
-               if (text.startsWith("lights ")) {
-                  val gName = text.substring(7).removeSuffix(" on").removeSuffix(" off")
-                  val g = groups.find { it.name.lowercase() equalsNc gName }
-                  if (g!=null) hueBridge.toggleBulbGroup(g.id).thenRefresh()
-               }
-            }
-            it.handlers += speechHandlers
-         }
+      hueBridge?.init()
+      hueBridge?.bulbsAndGroups().net { it?.first.orEmpty() to it?.second.orEmpty() }.net { (bulbs, groups) ->
          groupsPane.children setTo groups.map { group ->
             hueBulbGroupCells.getOrPut(group.id) {
                HueIcon(IconFA.LIGHTBULB_ALT, 40.0, group).run {
                   styleclass("hue-group-icon")
 
-                  fun toggleBulbGrouo() = hueBridge.toggleBulbGroup(group.id).thenRefresh()
-                  fun deleteBulbGrouo() = hueBridge.deleteGroup(group.id).thenRefresh()
+                  fun toggleBulbGrouo() = hueBridge?.toggleBulbGroup(group.id).thenRefresh()
+                  fun deleteBulbGrouo() = hueBridge?.deleteGroup(group.id).thenRefresh()
                   fun focusBulbGroup() {
                      unfocusSensor()
                      unfocusBulb()
@@ -560,7 +377,7 @@ class Hue(widget: Widget): SimpleController(widget) {
          }
          groupsPane.children += Icon(IconFA.PLUS).onClickDo {
             scope.launch(FX) {
-               val bulbsAll = hueBridge.bulbs()
+               val bulbsAll = hueBridge?.bulbs().orEmpty()
                object: ConfigurableBase<Any?>(), Validated {
                   var name by c("")
                   var type by c(Zone)
@@ -575,7 +392,7 @@ class Hue(widget: Widget): SimpleController(widget) {
                      else -> Try.ok()
                   }
                }.configure("Create group") {
-                  hueBridge.createGroup(it.materialize()).thenRefresh()
+                  hueBridge?.createGroup(it.materialize()).thenRefresh()
                }
             }
          }
@@ -589,22 +406,22 @@ class Hue(widget: Widget): SimpleController(widget) {
                      object: ConfigurableBase<Any?>() {
                         val name by cv(hue.name).lengthMax(32).def("Name")
                      }.configure("Rename bulb") {
-                        hueBridge.renameBulb(bulb.id, it.name.value).thenRefresh()
+                        hueBridge?.renameBulb(bulb.id, it.name.value).thenRefresh()
                      }
                   }
                   fun changePowerOn() {
                      object: ConfigurableBase<Any?>(), Validated {
                         val powerOn by cv(hue.confPowerOn ?: unknown).def("Behavior")
                         override fun isValid() = when {
-                           !hueBridge.apiVersion.isAtLeast(1, 28) -> Try.error("Unsupported before apiVersion 1.28. Please update bridge.")
+                           !hueBridge!!.apiVersion.isAtLeast(1, 28) -> Try.error("Unsupported before apiVersion 1.28. Please update bridge.")
                            powerOn.value == custom || powerOn.value == unknown -> Try.error("Value not supported")
                            else -> Try.ok()
                         }
                      }.configure("Change power on behavior") {
-                        hueBridge.changePowerOn(bulb.id, it.powerOn.value).thenRefresh()
+                        hueBridge?.changePowerOn(bulb.id, it.powerOn.value).thenRefresh()
                      }
                   }
-                  fun toggleBulb() = hueBridge.toggleBulb(bulb.id).thenRefresh()
+                  fun toggleBulb() = hueBridge?.toggleBulb(bulb.id).thenRefresh()
                   fun focusBulb() {
                      unfocusSensor()
                      unfocusBulbGroup()
@@ -645,7 +462,7 @@ class Hue(widget: Widget): SimpleController(widget) {
             }
          }
       }
-      hueBridge.scenes().net { scenes ->
+      hueBridge?.scenes().orEmpty().net { scenes ->
          scenesPane.children setTo scenes.map { scene ->
             fun focusScene() {
                unfocusSensor()
@@ -661,12 +478,12 @@ class Hue(widget: Widget): SimpleController(widget) {
                onEventDown(KEY_PRESSED, SPACE) { focusScene() }
                onEventDown(MOUSE_CLICKED, PRIMARY) {
                   if (it.clickCount==1) focusScene()
-                  if (it.clickCount==2) hueBridge.applyScene(scene)
+                  if (it.clickCount==2) hueBridge?.applyScene(scene)
                }
                onContextMenuRequested = EventHandler {
                   ContextMenu().dsl {
                      item("delete") {
-                        hueBridge.deleteScene(scene.id).thenRefresh()
+                        hueBridge?.deleteScene(scene.id).thenRefresh()
                      }
                   }.show(this, RIGHT, 0.0, 0.0)
                }
@@ -674,7 +491,7 @@ class Hue(widget: Widget): SimpleController(widget) {
          }
          scenesPane.children += Icon(IconFA.PLUS).onClickDo {
             scope.launch(FX) {
-               val bulbsAll = hueBridge.bulbs()
+               val bulbsAll = hueBridge?.bulbs().orEmpty()
                object: ConfigurableBase<Any?>(), Validated {
                   var name by c("")
                   val bulbs by cCheckList(*bulbsAll.toTypedArray()).uiConverterElement { it.name }
@@ -687,12 +504,12 @@ class Hue(widget: Widget): SimpleController(widget) {
                      else -> Try.ok()
                   }
                }.configure("Create scene") {
-                  hueBridge.createScene(it.materialize()).asFut().thenRefresh()
+                  hueBridge?.createScene(it.materialize())?.asFut().thenRefresh()
                }
             }
          }
       }
-      hueBridge.sensors().net { scenes ->
+      hueBridge?.sensors().orEmpty().net { scenes ->
          sensorsPane.children setTo scenes.map { sensor ->
             fun focusSensor() {
                unfocusSensor()
@@ -749,23 +566,6 @@ class Hue(widget: Widget): SimpleController(widget) {
       private var hueBridgeApiKey by appProperty("")
       private var hueBridgeIp by appProperty("")
 
-      fun JsObject.withoutNullValues() = JsObject(value.filter { it.value !is JsNull })
-      fun JsObject.withoutType() = JsObject(value.filter { it.key!="_type" })
-      fun Any?.toJson() = APP.serializerJson.json.toJsonValue(this).let {
-         when (it) {
-            is JsObject -> it.withoutType()
-            else -> it
-         }
-      }
-
-      fun String.parseToJson() = APP.serializerJson.json.ast(this).orThrow
-      inline fun <reified T> JsValue.to() = APP.serializerJson.json.fromJsonValue<T>(this).orThrow
-
-      suspend fun HttpClient.getText(url: String, block: HttpRequestBuilder.() -> Unit = {}) = get(url, block).bodyAsText()
-      suspend fun HttpClient.putText(url: String, block: HttpRequestBuilder.() -> Unit = {}) = put(url, block).bodyAsText()
-      suspend fun HttpClient.postText(url: String, block: HttpRequestBuilder.() -> Unit = {}) = post(url, block).bodyAsText()
-      suspend fun HttpClient.deleteText(url: String, block: HttpRequestBuilder.() -> Unit = {}) = delete(url, block).bodyAsText()
-
       fun drawGradientCir(radius: Int): Image {
          return WritableImage(2*radius, 2*radius).apply {
             val c = (radius).toDouble()
@@ -797,120 +597,4 @@ class Hue(widget: Widget): SimpleController(widget) {
          }
       }
    }
-}
-
-typealias HueBulbId = String
-typealias HueGroupId = String
-typealias HueSceneId = String
-typealias HueSensorId = String
-typealias HueMap = Map<String,Any?>
-
-class HueIcon<T>(i: GlyphIcons?, size: Double, var hue: T): Icon(i, size)
-class HueCell<T>(val node: Node, val icon: HueIcon<T>)
-class HueCellNode(icon: Icon, name: String): VBox(5.emScaled) {
-   var name: String = name
-      set(value) {
-         field = value
-         lookupId<Label>("nameLabel").text = value
-      }
-
-   init {
-      prefSize = 90.emScaled x 80.emScaled
-      alignment = TOP_CENTER
-      lay += icon.apply {
-         focusOwner.value = this@HueCellNode
-      }
-      lay += label(name) {
-         id = "nameLabel"
-         isWrapText = true
-         textAlignment = TextAlignment.CENTER
-      }
-   }
-}
-
-data class HueBulbStateEditOn(val on: Boolean)
-data class HueBulbStateEditLight(val bri: Int?, val hue: Int?, val sat: Int?)
-data class HueBulbState(val on: Boolean, val bri: Int, val hue: Int, val sat: Int, val reachable: Boolean)
-data class HueBulb(val id: HueBulbId = "", val name: String, val productname: String, val state: HueBulbState, val config: HueMap) {
-   val confPowerOn get() = config["startup"]?.asIf<HueMap>()?.get("mode")?.asIs<String>()?.net(HueBulbConfPowerOn::valueOf)
-}
-data class HueGroupState(val all_on: Boolean, val any_on: Boolean)
-data class HueGroup(val id: HueGroupId = "", val name: String, val lights: List<HueBulbId>, val state: HueGroupState, val type: String?)
-data class HueGroupCreate(val name: String, val type: String, val lights: List<HueBulbId>)
-data class HueScene(val id: HueSceneId = "", val name: String, val lights: List<HueBulbId>)
-data class HueSceneCreate(val name: String, val type: HueSceneType, val group: HueGroupId?, val lights: List<HueBulbId>?, val recycle: Boolean = false) {
-   constructor(name: String, lights: List<HueBulb>): this(name, LightScene, null, lights.map { it.id })
-   constructor(name: String, group: HueGroupId): this(name, GroupScene, group, null)
-}
-data class HueSensor(val id: HueSensorId = "", val name: String, val type: String, val state: HueMap, val config: HueMap) {
-   val stateTemperature get() = state["temperature"]?.asIf<Number>()?.net { it.toDouble()/100 }
-   val statePresence get() = state["presence"]?.asIf<Boolean>()
-   val stateDaylight get() = state["daylight"]?.asIf<Boolean>()
-   val icon: GlyphIcons get() = when (type) {
-      "ZLLTemperature" -> IconMD.THERMOMETER
-      "ZLLPresence" -> IconOC.BROADCAST
-      "ZLLLightLevel" -> IconWH.WU_CLEAR
-      "Daylight" -> IconMA.TIMELAPSE
-      else -> IconMA.SETTINGS_INPUT_ANTENNA
-   }
-}
-
-enum class HueBulbConfPowerOn(val value: String, override val infoUi: String): InfoUi {
-   custom("Custom", "Custom settings defined in custom settings. Will be automatically set when providing “customsettings”. Not available for “On/Off Light”."),
-   lastonstate("Last on state", "Light keeps the setting when power failed. If light was off it returns to the last on state."),
-   powerfail("Powerfail", "Light keeps the setting when power failed. If light was off it stays off."),
-   safety("Safety", "Lights go back to Philips “bright light” safety setting (100% brightness @ 2700K)."),
-   unknown("Unknown", "Custom setting is not supported.")
-}
-enum class HueSceneType { LightScene, GroupScene }
-enum class HueGroupType(val value: String, val since: Double, override val infoUi: String): InfoUi {
-   Luminaire(
-      "Luminaire",
-      1.4,
-      "" +
-         "Multisource luminaire group.\n\nA lighting installation of default groupings of hue " +
-         "lights. The bridge will pre-install these groups for ease of use. This type cannot be created manually. " +
-         "Also, a light can only be in a maximum of one luminaire group. See multisource luminaires for more info."
-   ),
-   Lightsource(
-      "Lightsource",
-      1.4,
-      "" +
-         "LightSource group.\n\nA group of lights which is created by the bridge based on " +
-         "multisource luminaire attributes of Zigbee light resource."
-   ),
-   LightGroup(
-      "LightGroup",
-      1.4,
-      "" +
-         "LightGroup group.\n\nA group of lights that can be controlled together. This is the " +
-         "default group type that the bridge generates for user created groups. Default type when no type is given on creation."
-   ),
-   Room(
-      "Room",
-      1.11,
-      "" +
-         "Room.\n\nA group of lights that are physically located in the same place in the house. Rooms " +
-         "behave similar as light groups, except: (1) A room can be empty and contain 0 lights, (2) a light is only " +
-         "allowed in one room and (3) a room isn’t automatically deleted when all lights in that room are deleted."
-   ),
-   Entertainment(
-      "Entertainment",
-      1.22,
-      "" +
-         "Represents an entertainment setup.\n\n" +
-         "Entertainment group describe a group of lights that are used in an entertainment setup. Locations describe " +
-         "the relative position of the lights in an entertainment setup. E.g. for TV the position is relative to the " +
-         "TV. Can be used to configure streaming sessions.\n\n" +
-         "Entertainment group behave in a similar way as light groups, with the exception: it can be empty and contain " +
-         "0 lights. The group is also not automatically recycled when lights are deleted. The group of lights can be " +
-         "controlled together as in LightGroup."
-   ),
-   Zone(
-      "Zone",
-      1.30,
-      "" +
-         "Zone.\n\nZones describe a group of lights that can be controlled together. Zones can be empty " +
-         "and contain 0 lights. A light is allowed to be in multiple zones."
-   )
 }